The give project is control flow graph generator using antlrv4, python . target language is Golanguage.
The given project works on making a control flow graph of Go language using antlrv4, python and visualise it using graph viz. It used the generic antlrv4 go grammar https://github.com/antlr/grammars-v4/tree/master/golang.
it works on basic programs and for basic statements like if-else, nested-if-else, for-loops and nested loops.